+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 
 - Added "no-operation" flag which prevents commands such as `bumpy increment` to persist changes
+- Added support for versions which use comma as a delimiter (e.g. Visual Studio C++ resource files contain versions such as "1,0,0,0")
 
 ## [0.9.0] - 2018-02-09